backend/requirements.txt
Flask
flask-cors
mysql-connector-python
face_recognition
opencv-python
numpy
python-dotenv
